Streamline and simplify macro definitions
The form `\aSomething` is used to denote a variable of a certain
category, not an operation on things. So context extension is simply
`\ctxtExtend`, etc.

`\aBigRule` is not terribly useful: if we need 3 premises (which we
will eventually need), we would have to introduce `\aBiggerRule`,
etc. It's ok to use `\aRule` and put the line breaks in the premises.